#364 Create VeLa: implemented function return type checking

+++ trunk/src/org/aavso/tools/vstar/util/locale/NumberParser.java
@@ -54,7 +54,7 @@ public class NumberParser {
 
 		switch (operand.getType()) {
 		case INTEGER:
-		case DOUBLE:
+		case REAL:
 			break;
 			
 		default:
